% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{MakeNGSfilter}
\alias{MakeNGSfilter}
\title{MakeNGSfilter}
\usage{
MakeNGSfilter()
}
\description{
Produces NGSfilter files
}
\details{
This Shiny app allows users to produce NGSfilter files from three input files:
\itemize{
\item a 'comments' file: a tab-separated table that lists all samples and controls and their characteristics;
\item a 'PCR plates design' file: a tab-separated file that represents the PCR plates design;
\item a 'tags plates design' file: a .xlsx files that represents the distribution of sample tags in the PCR plates context.
}

All input files are checked to avoid bugs when used in downstream programs and functions.

The output files can be used as input of the \code{ngsfilter} program of the OBITools (Boyer et al. 2016) and the \code{obifiles_to_metabarlist} function of the metabaR R package (Zinger et al. 2021).
}
\examples{
\donttest{
MakeNGSfilter() # opens the shiny app in RStudio viewer or web navigator
}

}
\references{
Boyer, F., Mercier, C., Bonin, A., Bras, Y. L., Taberlet, P., & Coissac, E. (2016). obitools : A unix-inspired software package for DNA metabarcoding. Molecular Ecology Resources, 16(1), 176‑182. https://doi.org/10.1111/1755-0998.12428

Zinger, L., Lionnet, C., Benoiston, A.-S., Donald, J., Mercier, C., & Boyer, F. (2021). metabaR : An r package for the evaluation and improvement of DNA metabarcoding data quality. Methods in Ecology and Evolution, 12(4), 586‑592. https://doi.org/10.1111/2041-210X.13552
}
\author{
Anne-Sophie Benoiston
}
